Refactor and add unit test (and more soon)
This creates a separate static library target with some shared code
moved (parser, extensions) into it. This is required to make a unit
testing target, as app extensions can not be unit tested. Some other
code that the root application also uses for SwiftUI is in a shared
target memebership instead of in the library due to linker magic.

The parser class has also been refactored so HTML can be provided
directly to simplify unit testing, as well as clean up some debugging
stuff.

I'm not entirely pleased with this (it seems pretty noisy), but it does
seem to work without introducing additional files in the bundles, and
allows for future expansion.